+++ src/main/java/com/bc/pmpheep/back/service/WriterRoleServiceImpl.java
@@ -11,6 +11,8 @@ import com.bc.pmpheep.back.po.WriterPermission;
 import com.bc.pmpheep.back.po.WriterRole;
 import com.bc.pmpheep.back.po.WriterRolePermission;
 import com.bc.pmpheep.back.po.WriterUserRole;
+import com.bc.pmpheep.service.exception.CheckedExceptionBusiness;
+import com.bc.pmpheep.service.exception.CheckedExceptionResult;
 import com.bc.pmpheep.service.exception.CheckedServiceException;
 
 /**
@@ -27,25 +29,41 @@ public class WriterRoleServiceImpl implements WriterRoleService {
 
     @Override
     public WriterRole add(WriterRole writerRole) throws CheckedServiceException {
+        if (null == writerRole) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色属性为空时禁止新增角色");
+        }
         roleDao.add(writerRole);
         return writerRole;
     }
 
     @Override
-    public Integer delete(int id) throws CheckedServiceException {
+    public Integer delete(Long id) throws CheckedServiceException {
+        if (null == id) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID为空时禁止删除");
+        }
         return roleDao.delete(id);
     }
 
     @Transactional
     @Override
-    public void deleteRoleAndResource(List<Integer> ids) throws CheckedServiceException {
+    public void deleteRoleAndResource(List<Long> ids) throws CheckedServiceException {
+        if (0 == ids.size()) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID为空时禁止删除");
+        }
         roleDao.batchDelete(ids);
         roleDao.batchDeleteRoleResource(ids);
 
     }
 
     @Override
-    public WriterRole get(int id) throws CheckedServiceException {
+    public WriterRole get(Long id) throws CheckedServiceException {
+        if (null == id) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID为空时禁止查询");
+        }
         return roleDao.get(id);
     }
 
@@ -56,6 +74,10 @@ public class WriterRoleServiceImpl implements WriterRoleService {
 
     @Override
     public Integer update(WriterRole role) throws CheckedServiceException {
+        if (null == role.getId()) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID为空时禁止更新");
+        }
         return roleDao.update(role);
     }
 
@@ -65,53 +87,89 @@ public class WriterRoleServiceImpl implements WriterRoleService {
     }
 
     @Override
-    public WriterUserRole getUserRole(int uid, int roleId) throws CheckedServiceException {
+    public WriterUserRole getUserRole(Long uid, Long roleId) throws CheckedServiceException {
+        if (null == uid || null == roleId) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID或用户ID为空时禁止查询");
+        }
         return roleDao.getUserRole(uid, roleId);
     }
 
     @Override
-    public void addUserRole(int uid, int roleId) throws CheckedServiceException {
+    public void addUserRole(Long uid, Long roleId) throws CheckedServiceException {
+        if (null == uid || null == roleId) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID或用户ID为空时禁止新增");
+        }
         roleDao.addUserRole(uid, roleId);
 
     }
 
     @Override
-    public void deleteUserRole(int uid, int roleId) throws CheckedServiceException {
+    public void deleteUserRole(Long uid, Long roleId) throws CheckedServiceException {
+        if (null == uid || null == roleId) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID或用户ID为空时禁止删除");
+        }
         roleDao.deleteUserRole(uid, roleId);
 
     }
 
     @Override
-    public void deleteUserRoles(int uid) throws CheckedServiceException {
+    public void deleteUserRoles(Long uid) throws CheckedServiceException {
+        if (null == uid) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.USER_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "用户ID为空时禁止删除");
+        }
         roleDao.deleteUserRoles(uid);
 
     }
 
     @Override
-    public List<WriterPermission> getListRoleResource(int roleId) throws CheckedServiceException {
+    public List<WriterPermission> getListRoleResource(Long roleId) throws CheckedServiceException {
+        if (null == roleId) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.USER_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID为空时禁止查询");
+        }
         return roleDao.getListRoleResource(roleId);
     }
 
     @Override
-    public void addRoleResource(int roleId, int resId) throws CheckedServiceException {
+    public void addRoleResource(Long roleId, Long resId) throws CheckedServiceException {
+        if (null == roleId || null == resId) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID或资源ID为空时禁止新增");
+        }
         roleDao.addRoleResource(roleId, resId);
 
     }
 
     @Override
-    public void deleteRoleResource(int roleId, int resId) throws CheckedServiceException {
+    public void deleteRoleResource(Long roleId, Long resId) throws CheckedServiceException {
+        if (null == roleId || null == resId) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID或资源ID为空时禁止删除");
+        }
         roleDao.deleteRoleResource(roleId, resId);
 
     }
 
     @Override
-    public WriterRolePermission getResourceRole(int roleId, int resId)
+    public WriterRolePermission getResourceRole(Long roleId, Long resId)
     throws CheckedServiceException {
+        if (null == roleId || null == resId) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "角色ID或资源ID为空时禁止查询");
+        }
         return roleDao.getResourceRole(roleId, resId);
     }
 
     @Override
-    public Integer deleteRoleAndUser(List<Integer> ids) throws CheckedServiceException {
+    public Integer deleteRoleAndUser(List<Long> ids) throws CheckedServiceException {
+        if (0 == ids.size()) {
+            throw new CheckedServiceException(CheckedExceptionBusiness.ROLE_MANAGEMENT,
+                                              CheckedExceptionResult.NULL_PARAM, "用户ID为空时禁止删除");
+        }
         return roleDao.deleteRoleAndUser(ids);
     }
